Add: Clientから送られてきたJSONによってハンドリングするようにしました
